+++ CHANGELOG.md
@@ -7,7 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- Add auto-accept plan mode patch (#464) - @irdbl
 - Add a fallback for WASMagic when it's not available (#399) - @signadou
 - Add opusplan[1m] model alias for 1M context support (#404) - @mike1858
 - Add MCP startup optimization settings (#407) - @mike1858
@@ -27,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix `subagentModels` patch to not error when nothing was changed (#456) - @bl-ue
 - Enable the builtin `/remember` skill (#457) - @bl-ue
 - AGENTS.md support for Claude Code (#459) - @bl-ue
+- Add auto-accept plan mode patch (#464) - @irdbl
+- Add `--config-url` flag to fetch configuration from a URL (#465) - @bl-ue & @basekevin
 
 ## [v3.4.0](https://github.com/Piebald-AI/tweakcc/releases/tag/v3.4.0) - 2026-01-18